Develop T3b-1: the masks header stops being three headers in a trenchcoat
The survey called the dt_masks_functions_t vtable the hardest knot in
src/develop: one per-shape table carrying pipeline, GUI and data function
pointers, forcing masks.h to define dt_masks_form_gui_t and include
widgets/draw.h -- GTK, ultimately -- for every consumer that only wanted
get_mask(). The prescribed fix was a parallel GUI vtable plus a registry.

Measurement found something better: EVERY member of the table is dispatched
from inside src/develop/masks/ and nowhere else. The table is an implementation
detail that happened to be public. So there is no split and no registry:

- develop/masks/masks_functions.h (internal): the full table definition and the
  six per-shape extern tables. Only the shape files and the masks core include
  it. Public code holds the table as an opaque pointer behind a forward
  declaration.
- develop/masks_gui.h: dt_masks_form_gui_t, the hit-testing/selection inlines,
  bezier/projection/draw-preview helpers, shape buttons, event dispatchers,
  creation flows, menus, dialogs, the interaction API -- and the widgets/draw.h
  include, which is theirs.
- develop/masks.h keeps the data model (types, nodes, forms, refcounting), the
  rasterisation entry points, persistence, groups and dynbuf -- and now contains
  ZERO toolkit types.

Four inline dispatchers stood in the way -- get_mask/get_mask_roi in masks.h,
the two bezier togglers in masks_gui.h -- and are de-inlined: a per-buffer call
is not a per-pixel cost, so the inline bought nothing but the table's publicity.

Twenty-two riders surfaced, the largest supply chain cut yet: masks.h ->
draw.h -> widget_settings.h had been feeding DT_GUI_BOX_SPACING,
DT_PIXEL_APPLY_DPI, dt_gui_freeze_*, dt_modifier_is and friends to fourteen
files, and the masks GUI API itself to seven IOPs. Each now names its supplier.

One rider was not a rider but a resident: control/jobs/import_jobs.c builds the
discarded-files recap dialog -- a full GTK dialog in a control-layer job file,
which the explicit include would have turned into a counted upward edge. The
dialog moves to gui/import.c behind a handler; the job calls it on the worker
thread and the HANDLER hops to the GUI main loop, because g_main_context_invoke
is toolkit machinery and not a job's business (maintainer's observation). The
handler owns freeing the enumerator params, exactly as the popup did; headless,
the job frees them itself. import_jobs.c thereby loses all three of its upward
includes (gui/application.h, widgets/widget_style.h, and the would-be
widget_settings.h) -- layering violations FALL 189 -> 187, locked in.

Release, Debug (-Werror), nofeatures all build.
