+++ src/darktable.c
@@ -143,6 +143,7 @@
 #include "metadata/metadata.h"
 #include "common/image_notify.h"
 #include "develop/dev_history_gui.h"
+#include "gui/import.h"
 #include "develop/pipeline_notify.h"
 #include "history/notify.h"
 #include "history/presets.h"
@@ -1515,6 +1516,7 @@ int dt_init(int argc, char *argv[], const gboolean init_gui, const gboolean load
    * edit a tag before it does. */
   dt_metadata_set_tags_changed_handler(_metadata_tags_changed);
   dt_dev_history_gui_init();
+  dt_gui_import_init_handlers();
   dt_metadata_set_geotags_changed_handler(_metadata_geotags_changed);
   dt_image_notify_set_imported_handler(_image_imported);
   dt_pipeline_set_message_handler(_pipeline_message);
