+++ src/develop/masks/brush.c
@@ -45,6 +45,8 @@
 #include "common/conf.h"
 #include "develop/imageop.h"
 #include "develop/masks.h"
+#include "develop/masks_gui.h"
+#include "develop/masks/masks_functions.h"
 #include "math/openmp_maths.h"
 #include "gui/actions/menu.h"
