+++ src/develop/masks/polygon.c
@@ -50,6 +50,8 @@
 #include "common/conf.h"
 #include "develop/imageop.h"
 #include "develop/masks.h"
+#include "develop/masks_gui.h"
+#include "develop/masks/masks_functions.h"
 #include "math/openmp_maths.h"
 #include "gui/actions/menu.h"
 #include <assert.h>
