+++ src/iop/retouch.c
@@ -68,6 +68,7 @@
 #include "develop/imageop_math.h"
 #include "develop/imageop_gui.h"
 #include "develop/masks.h"
+#include "develop/masks_gui.h"
 #include "develop/tiling.h"
 #include "gui/actions/menu.h"
 #include "iop/iop_api.h"
